grpc auth: add file based grpc credentials extension (#3438)
Adds a file based grpc credentials extension. See issue #3392 for more details.

Risk Level: Low: extension for grpc credentials loaded by explicit configuration options

Testing: tests included in PR

Docs Changes: Inline docs via comments and proto docs

Release Notes: N/A

Fixes #3392
