+++ source/common/grpc/google_grpc_creds_impl.cc
@@ -4,8 +4,6 @@
 #include "envoy/grpc/google_grpc_creds.h"
 #include "envoy/registry/registry.h"
 
-#include "common/config/datasource.h"
-
 namespace Envoy {
 namespace Grpc {
 
@@ -13,15 +11,6 @@ namespace Grpc {
  * TODO: Create GoogleGrpcCredentialsFactory for each built-in credential type defined in Google
  * gRPC.
  */
-
-grpc::SslCredentialsOptions buildSslOptionsFromConfig(
-    const envoy::api::v2::core::GrpcService::GoogleGrpc::SslCredentials& ssl_config) {
-  return {
-      .pem_root_certs = Config::DataSource::read(ssl_config.root_certs(), true),
-      .pem_private_key = Config::DataSource::read(ssl_config.private_key(), true),
-      .pem_cert_chain = Config::DataSource::read(ssl_config.cert_chain(), true),
-  };
-}
 
 /**
  * Default implementation of Google Grpc Credentials Factory
@@ -35,14 +24,7 @@ class DefaultGoogleGrpcCredentialsFactory : public GoogleGrpcCredentialsFactory 
 public:
   std::shared_ptr<grpc::ChannelCredentials>
   getChannelCredentials(const envoy::api::v2::core::GrpcService& grpc_service_config) override {
-    const auto& google_grpc = grpc_service_config.google_grpc();
-    std::shared_ptr<grpc::ChannelCredentials> creds = grpc::InsecureChannelCredentials();
-    if (google_grpc.has_channel_credentials() &&
-        google_grpc.channel_credentials().has_ssl_credentials()) {
-      return grpc::SslCredentials(
-          buildSslOptionsFromConfig(google_grpc.channel_credentials().ssl_credentials()));
-    }
-    return creds;
+    return defaultSslChannelCredentials(grpc_service_config);
   }
 
   std::string name() const override { return "envoy.grpc_credentials.default"; }
