+++ source/extensions/extensions_build_config.bzl
@@ -8,6 +8,12 @@ EXTENSIONS = {
     "envoy.access_loggers.http_grpc":                   "//source/extensions/access_loggers/http_grpc:config",
 
     #
+    # gRPC Credentials Plugins
+    #
+
+    "envoy.grpc_credentials.file_based_metadata":      "//source/extensions/grpc_credentials/file_based_metadata:config",
+
+    #
     # Health checkers
     #
