+++ source/extensions/grpc_credentials/well_known_names.h
@@ -14,6 +14,8 @@ class GrpcCredentialsNameValues {
 public:
   // Access Token Example.
   const std::string ACCESS_TOKEN_EXAMPLE = "envoy.grpc_credentials.access_token_example";
+  // File Based Metadata credentials
+  const std::string FILE_BASED_METADATA = "envoy.grpc_credentials.file_based_metadata";
 };
 
 typedef ConstSingleton<GrpcCredentialsNameValues> GrpcCredentialsNames;
